Rename cast_ref_to_mut lint to invalid_reference_casting
(cherry picked from commit 3dbbf23e29516218863bda29d2983bd503e6b7fd)
