+++ tests/ui/lint/reference_casting.rs
